requirements.txt

fastapi
uvicorn
pydantic[dotenv]
python-multipart
requests
python-docx
pymupdf
sentence-transformers
faiss-cpu
huggingface_hub
hf_transfer
hf_xet
